Util: Separate out Util-IO from Util
In order to allow building on embedded platforms (without any
TCTIs) the IO functions from util have been separated out into
their own library
