+++ src/tss2-tcti/tcti-spi-ltt2go.c
@@ -18,7 +18,6 @@
 #include "tss2_tcti_spi_ltt2go.h"
 #include "tcti-spi-helper.h"
 #include "tss2_mu.h"
-#include "util/io.h"
 #define LOGMODULE tcti
 #include "util/log.h"
