+++ test/unit/log.c
@@ -18,7 +18,7 @@
 
 #include "tss2_tpm2_types.h"
 
-#include "util/io.h"
+#include "util-io/io.h"
 #define LOGMODULE test
 #include "util/log.h"
